add: A new typeclass [Nat] to abstract away natural numbers
The current status in Coq is that you can use either [nat] or [N] to
deal with (infinite) natural numbers, but as far as I know this is not
possible to write eg. a recursive function that will work with
both. The [Nat] typeclass is an attempt to solve this issue.