refactor: Make coq-prelude less opinionated
coq-prelude was initially an attempt to provide a more Haskell-flavor
API to Coq development. This has proven to be less effective than
anticipated in practice.

This patch is a first important step toward making coq-prelude less
opinionated. In particular:

  - We reuse existing type as much as possible (for instance, no more
    [Either], we use [sum] instead)
  - The monad notations form a compromise between the Haskell
    do-notation and OCaml 4.08 new “monadic” operators
  - We remove the [Integer] libraries

With this patch, the minimum Coq version to use coq-prelude is
coq.8.10.0.